# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  entities.UserUpdateRequest:
    properties:
      firstname:
        type: string
      gender:
        type: string
      icon_url:
        type: string
      lastname:
        type: string
      middlename:
        type: string
      phone:
        type: string
      user_id:
        type: integer
    required:
    - firstname
    - gender
    - icon_url
    - lastname
    - middlename
    - phone
    - user_id
    type: object
  userv1.UpdateInfoResponse:
    properties:
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Gatewate for orbit of success services
  title: API Gatewate
  version: "1.0"
paths:
...
  /user:
    put:
      consumes:
      - application/json
      description: Update user info
      operationId: Update user info
      parameters:
      - description: info
        in: body
        name: info
        schema:
          $ref: '#/definitions/entities.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/userv1.UpdateInfoResponse'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
        "503":
          description: Service Unavailable
      summary: Update user info
      tags:
      - User
swagger: "2.0"